frontend_with_backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
